Lazy load modals and page
