+++ frontend/package.json
@@ -50,9 +50,11 @@
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.8",
+    "@loadable/component": "^5.16.3",
     "@mui/icons-material": "^5.15.10",
     "@mui/material": "^5.15.10",
     "@types/jest": "^29.5.12",
+    "@types/loadable__component": "^5.13.8",
     "@types/lodash": "^4.14.202",
     "@types/node": "^20.11.17",
     "@types/react": "^18.2.55",
