+++ frontend/src/component/modal/MarkNewTagEditorModal.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useContext, useMemo, useState } from 'react';
 import Dialog from '@mui/material/Dialog';
 import {
     DialogActions,
@@ -9,36 +9,35 @@ import {
 import Stack from '@mui/material/Stack';
 import { Trans, useTranslation } from 'react-i18next';
 import NiceModal, { muiDialog, useModal } from '@ebay/nice-modal-react';
-import { logError } from '../util';
-import { CancelButton, SaveButton } from './Buttons';
-import { UserSettings } from '../api';
-import { sortedUniq } from 'lodash';
+import { logError } from '../../util.ts';
+import { CancelButton, SaveButton } from '../Buttons.tsx';
+import { SettingsContext } from '../../context/SettingsContext.ts';
 
-interface KickTagEditorProps {
-    setNewSettings: React.Dispatch<React.SetStateAction<UserSettings>>;
+interface MarkNewTagEditorProps {
+    steam_id: string;
+    onMarkAs: (steamId: string, attrs: string[]) => Promise<void>;
 }
 
-export const SettingsKickTagEditor = NiceModal.create<KickTagEditorProps>(
-    ({ setNewSettings }) => {
+export const MarkNewTagEditorModal = NiceModal.create<MarkNewTagEditorProps>(
+    ({ steam_id, onMarkAs }) => {
         const [tag, setTag] = useState<string>('');
         const { t } = useTranslation();
         const modal = useModal();
+        const { settings, setSettings } = useContext(SettingsContext);
 
-        const onSaveTag = useCallback(async () => {
+        const onSaveMarkWithNewTag = useCallback(async () => {
             try {
-                setNewSettings((prevState) => {
-                    return {
-                        ...prevState,
-                        kick_tags: sortedUniq([...prevState.kick_tags, tag])
-                    };
+                await onMarkAs(steam_id, [tag]);
+                setSettings({
+                    ...settings,
+                    unique_tags: [...tag, ...settings.unique_tags]
                 });
-                setTag('');
             } catch (e) {
                 logError(`Error updating note: ${e}`);
             } finally {
                 await modal.hide();
             }
-        }, [setNewSettings, tag, modal]);
+        }, [onMarkAs, steam_id, tag, setSettings, settings, modal]);
 
         const validTag = useMemo(() => {
             return tag.length > 0 && !tag.match(/\s/);
@@ -47,14 +46,14 @@ export const SettingsKickTagEditor = NiceModal.create<KickTagEditorProps>(
         return (
             <Dialog fullWidth {...muiDialog(modal)}>
                 <DialogTitle>
-                    <Trans i18nKey={'new_kick_tag.title'} />
+                    <Trans i18nKey={'mark_new_tag.title'} />
                 </DialogTitle>
                 <DialogContent>
                     <Stack spacing={1} padding={0}>
                         <TextField
                             error={tag.length > 0 && !validTag}
                             id="new-tag-editor-field"
-                            label={t('new_kick_tag.tag')}
+                            label={t('mark_new_tag.tag')}
                             fullWidth
                             value={tag}
                             onChange={(evt) => {
@@ -65,9 +64,14 @@ export const SettingsKickTagEditor = NiceModal.create<KickTagEditorProps>(
                 </DialogContent>
                 <DialogActions>
                     <CancelButton onClick={modal.hide} />
-                    <SaveButton onClick={onSaveTag} disabled={!validTag} />
+                    <SaveButton
+                        onClick={onSaveMarkWithNewTag}
+                        disabled={!validTag}
+                    />
                 </DialogActions>
             </Dialog>
         );
     }
 );
+
+export default MarkNewTagEditorModal;
