+++ frontend/src/component/modal/SettingsKickTagEditorModal.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useContext, useMemo, useState } from 'react';
+import { useCallback, useMemo, useState } from 'react';
 import Dialog from '@mui/material/Dialog';
 import {
     DialogActions,
@@ -9,35 +9,36 @@ import {
 import Stack from '@mui/material/Stack';
 import { Trans, useTranslation } from 'react-i18next';
 import NiceModal, { muiDialog, useModal } from '@ebay/nice-modal-react';
-import { logError } from '../util';
-import { CancelButton, SaveButton } from './Buttons';
-import { SettingsContext } from '../context/SettingsContext';
+import { logError } from '../../util.ts';
+import { CancelButton, SaveButton } from '../Buttons.tsx';
+import { UserSettings } from '../../api.ts';
+import { sortedUniq } from 'lodash';
 
-interface MarkNewTagEditorProps {
-    steam_id: string;
-    onMarkAs: (steamId: string, attrs: string[]) => Promise<void>;
+interface KickTagEditorProps {
+    setNewSettings: React.Dispatch<React.SetStateAction<UserSettings>>;
 }
 
-export const MarkNewTagEditor = NiceModal.create<MarkNewTagEditorProps>(
-    ({ steam_id, onMarkAs }) => {
+export const SettingsKickTagEditorModal = NiceModal.create<KickTagEditorProps>(
+    ({ setNewSettings }) => {
         const [tag, setTag] = useState<string>('');
         const { t } = useTranslation();
         const modal = useModal();
-        const { settings, setSettings } = useContext(SettingsContext);
 
-        const onSaveMarkWithNewTag = useCallback(async () => {
+        const onSaveTag = useCallback(async () => {
             try {
-                await onMarkAs(steam_id, [tag]);
-                setSettings({
-                    ...settings,
-                    unique_tags: [...tag, ...settings.unique_tags]
+                setNewSettings((prevState) => {
+                    return {
+                        ...prevState,
+                        kick_tags: sortedUniq([...prevState.kick_tags, tag])
+                    };
                 });
+                setTag('');
             } catch (e) {
                 logError(`Error updating note: ${e}`);
             } finally {
                 await modal.hide();
             }
-        }, [onMarkAs, steam_id, tag, setSettings, settings, modal]);
+        }, [setNewSettings, tag, modal]);
 
         const validTag = useMemo(() => {
             return tag.length > 0 && !tag.match(/\s/);
@@ -46,14 +47,14 @@ export const MarkNewTagEditor = NiceModal.create<MarkNewTagEditorProps>(
         return (
             <Dialog fullWidth {...muiDialog(modal)}>
                 <DialogTitle>
-                    <Trans i18nKey={'mark_new_tag.title'} />
+                    <Trans i18nKey={'new_kick_tag.title'} />
                 </DialogTitle>
                 <DialogContent>
                     <Stack spacing={1} padding={0}>
                         <TextField
                             error={tag.length > 0 && !validTag}
                             id="new-tag-editor-field"
-                            label={t('mark_new_tag.tag')}
+                            label={t('new_kick_tag.tag')}
                             fullWidth
                             value={tag}
                             onChange={(evt) => {
@@ -64,12 +65,11 @@ export const MarkNewTagEditor = NiceModal.create<MarkNewTagEditorProps>(
                 </DialogContent>
                 <DialogActions>
                     <CancelButton onClick={modal.hide} />
-                    <SaveButton
-                        onClick={onSaveMarkWithNewTag}
-                        disabled={!validTag}
-                    />
+                    <SaveButton onClick={onSaveTag} disabled={!validTag} />
                 </DialogActions>
             </Dialog>
         );
     }
 );
+
+export default SettingsKickTagEditorModal;
